+++ hw09-Jdbc/src/main/java/ru/otus/orm/api/dao/UserDao.java
@@ -1,7 +1,7 @@
 package ru.otus.orm.api.dao;
 
-import ru.otus.orm.api.sessionmanager.SessionManager;
 import ru.otus.orm.model.User;
+import ru.otus.orm.api.sessionmanager.SessionManager;
 
 import java.util.Optional;
