+++ hw09-Jdbc/src/main/java/ru/otus/orm/jdbc/dao/UserDaoException.java
@@ -1,4 +1,4 @@
-package ru.otus.orm.api.dao;
+package ru.otus.orm.jdbc.dao;
 
 public class UserDaoException extends RuntimeException {
   public UserDaoException(Exception ex) {
