+++ hw09-Jdbc/src/main/java/ru/otus/orm/jdbc/dao/UserDaoJdbc.java
@@ -1,11 +1,12 @@
-package ru.otus.orm.api.dao;
+package ru.otus.orm.jdbc.dao;
 
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.otus.orm.DbExecutor;
+import ru.otus.orm.api.dao.UserDao;
 import ru.otus.orm.api.sessionmanager.SessionManager;
-import ru.otus.orm.api.sessionmanager.SessionManagerJdbc;
+import ru.otus.orm.jdbc.sessionmanager.SessionManagerJdbc;
 import ru.otus.orm.model.User;
 
 import java.sql.Connection;
